+++ spec/unit/generators/config_generator_spec.rb
@@ -20,6 +20,7 @@ RSpec.describe 'Generators' do
         { name: '/config/custom.yml', is_directory: false, size: template_file_size('config', 'custom.yml') },
         { name: '/config/default.yml', is_directory: false, size: template_file_size('config', 'default.yml') },
         { name: '/config/drivers', is_directory: true },
+        { name: '/config/drivers/appium.rb', is_directory: false, size: 724 },
         { name: '/config/drivers/browserstack.rb', is_directory: false, size: 1018 },
         { name: '/config/drivers/crossbrowsertesting.rb', is_directory: false, size: 1237 },
         { name: '/config/drivers/headless_chrome.rb', is_directory: false, size: 649 },
@@ -41,6 +42,7 @@ RSpec.describe 'Generators' do
       #{ColorizedString.new('Added').light_green} 'config/custom.yml' file
       #{ColorizedString.new('Added').light_green} 'config/capybara.rb' file
       #{ColorizedString.new('Added').light_green} 'config/default.yml' file
+      #{ColorizedString.new('Added').light_green} 'config/drivers/appium.rb' file
       #{ColorizedString.new('Added').light_green} 'config/drivers/browserstack.rb' file
       #{ColorizedString.new('Added').light_green} 'config/drivers/crossbrowsertesting.rb' file
       #{ColorizedString.new('Added').light_green} 'config/drivers/headless_chrome.rb' file